refactor : inputView, outputView 분리
